==> notebooks/single-cell/ustekinumab_cohort/03-CD4Teff_analysis/04-refine_annotations.ipynb <==
working_dir = "/data/projects/dschaub/ANCA-GN_transcriptomics"
data_dir = "/data/projects/dschaub/ANCA-GN_transcriptomics/data/single-cell/ustekinumab"

%load_ext autoreload
%autoreload 2

import sys
import os

os.chdir(working_dir)

import yaml
import anndata as ad
import matplotlib.pyplot as plt
import seaborn as sns
import mudata as md
import muon as mu
import numpy as np
import scanpy as sc
from matplotlib import gridspec
import scanpy.external as sce
from scipy import sparse
from matplotlib.colors import LinearSegmentedColormap
import matplotlib as mpl

from utils.utils import *
from utils.plotting import *

sc.settings.verbosity = 0
sc.set_figure_params(dpi=80)
sns.set(style="white")

path = os.path.join(data_dir, "ANCA_ustekinumab_4PK4PB_CD4Teff.h5mu")
anca_data = md.read_h5mu(path)
anca_data

anca_data.obs

relevant_obs = [
    "patient",
    "sample",
    "tissue",
    "case",
    "orig.ident",
    "cell_type",
    "RNA_snn_res.0.3",
]
anca_data.obs = anca_data.obs[relevant_obs]

mod = "RNA"
rna_data = ad.AnnData(
    anca_data.mod[mod].layers["counts"].copy(),
    # obs=anca_data.obs.copy(),
    var=anca_data.mod[mod].var[[]],
    layers={"counts": anca_data.mod[mod].layers["counts"].copy()},
    obsm={
        "X_harmony": anca_data.obsm["X_harmony"].copy(),
        "X_umap": anca_data.obsm["X_umap"].copy(),
    },
)

mod = "CITE"
cite_data = ad.AnnData(
    anca_data.mod[mod].layers["counts"].copy(),
    # obs=anca_data.obs.copy(),
    var=anca_data.mod[mod].var[[]],
    layers={"counts": anca_data.mod[mod].layers["counts"].copy()},
    obsm={
        "X_harmony": anca_data.obsm["X_harmony"].copy(),
        "X_umap": anca_data.obsm["X_umap"].copy(),
    },
)

rna_data.obs = anca_data.obs.copy()
cite_data.obs = anca_data.obs.copy()

rna_data.obs["RNA_snn_res.0.3"].unique().tolist()

rna_data.obs["cell_type_fine"] = rna_data.obs["RNA_snn_res.0.3"].replace(
    {
        "0": "TH1-like CD4+ EM",
        "1": "other CD4+ EM",
        "2": "TFH",
        "3": "TRM1",
        "4": "TRM17",
        "5": "TFH",
    }
)

anca_data_clean = md.MuData(
    {
        "rna": rna_data.copy(),
        "cite": cite_data.copy(),
    }
)

# save data
save_path = os.path.join(data_dir, "ANCA_ustekinumab_4PK4PB_CD4Teff_annotated.h5mu")
anca_data_clean.write_h5mu(save_path)
